curl.h: switch `CURL_HTTP_VERSION*` enums to long constants
It fixes tests 1539, and 2402, 2404 (for non-Secure Transport), on macOS
with the gcc compiler.

Also unignore these tests in GHA/macos for non-secure transport.

Ref: c349bd668c91f2484ae21c0f361ddf497143093c #14097 (issue 15.)
Ref: 7b0240c07799c28dc84272f9e38e1092ce4cc498 #16539
Ref: 2ec00372a1fc7f27cd3a6c43e29007400acfe2b6 #16482

Closes #16580

+++ tests/libtest/lib1901.c
@@ -70,7 +70,7 @@ CURLcode test(char *URL)
     easy_setopt(curl, CURLOPT_READFUNCTION, read_callback);
     easy_setopt(curl, CURLOPT_POST, 1L);
     easy_setopt(curl, CURLOPT_VERBOSE, 1L);
-    easy_setopt(curl, CURLOPT_HTTP_VERSION, (long)CURL_HTTP_VERSION_1_1);
+    easy_setopt(curl, CURLOPT_HTTP_VERSION, CURL_HTTP_VERSION_1_1);
     easy_setopt(curl, CURLOPT_URL, URL);
     easy_setopt(curl, CURLOPT_READDATA, NULL);
